refactor: move note creation logic to hook
